Code Updates
Removes whitespace in one of the files. Thanks @ValliantStorme for finding this problem and reporting it.
Switches comments to be in brackets instead of AR comments for a couple files. (this is just the start of a much bigger project and will take a very long time to finish)
Removes a set of duplicate codes found in one of the cheat files.